Move FromVoid and ItemRef to base
This will let us reuse them in dictionary.rs
